requirements.txt
cryptography>=2.7
Jinja2>=2.10.1
python-socketio>=4.1.0
plyvel>=1.1.0
pypng>=0.0.20
pillow>=6.1.0
pyzbar>=0.1.8
ecdsa>=0.14.1
aiohttp>=3.6.2
aiohttp-jinja2>=1.2.0
python-ndn>=0.2b1
PyYAML >= 5.1.2